feat(templates): Refactor template gallery to create lightweight version to export in package
