+++ apps/Standalone/src/templates/app/TemplatesStandard.tsx
@@ -58,7 +58,13 @@ export const TemplatesStandard = () => {
     theme: state.workflowLoader.theme,
     templatesView: state.workflowLoader.templatesView,
   }));
-  const { appId, hostingPlan, workflowName: existingWorkflowName, useEndpoint } = useSelector((state: RootState) => state.workflowLoader);
+  const {
+    appId,
+    hostingPlan,
+    workflowName: existingWorkflowName,
+    useEndpoint,
+    enableResourceSelection,
+  } = useSelector((state: RootState) => state.workflowLoader);
   const { data: workflowAppData } = useWorkflowApp(appId as string, hostingPlan);
   const canonicalLocation = useMemo(
     () => WorkflowUtility.convertToCanonicalFormat(workflowAppData?.location ?? 'westus'),
@@ -271,7 +277,7 @@ export const TemplatesStandard = () => {
         isCreateView={true}
         existingWorkflowName={existingWorkflowName}
         reload={reload}
-        enableResourceSelection={true}
+        enableResourceSelection={enableResourceSelection}
         viewTemplate={
           isSingleTemplateView
             ? {
