+++ libs/designer-ui/src/lib/templates/templates.less
@@ -94,10 +94,7 @@
 .msla-templates-filter-dropdown {
     margin: 10px;
     width: 200px;
-
-    .msal-template-filter-pill-name {
-        margin-right: 5px;
-    }
+    border-radius: 4px;
 }
 
 .msla-template-create-tabs {
