+++ libs/designer/src/lib/ui/templates/__tests__/templateCard.spec.tsx
@@ -7,6 +7,7 @@ import { screen } from '@testing-library/react';
 import { TemplateCard } from '../cards/templateCard';
 // biome-ignore lint/correctness/noUnusedImports: <explanation>
 import React from 'react';
+import { openQuickViewPanelView } from '../../../core/state/templates/panelSlice';
 
 describe('ui/templates/templatesDesigner', () => {
   let store: AppStore;
@@ -85,6 +86,7 @@ describe('ui/templates/templatesDesigner', () => {
           sortKey: 'a-to-z',
           connectors: undefined,
           detailFilters: {},
+          pageNum: 0,
         },
       },
     };
@@ -97,7 +99,7 @@ describe('ui/templates/templatesDesigner', () => {
     expect(store.getState().template.templateName).toBe('template1');
   });
 
-  it('Renders TemplateCard and Opens the right panel', async () => {
+  it('Renders TemplateCard and calls select handler on selection', async () => {
     minimalStoreData = {
       manifest: {
         availableTemplateNames: ['template1', 'template2', 'template3'],
@@ -110,12 +112,15 @@ describe('ui/templates/templatesDesigner', () => {
           sortKey: 'a-to-z',
           connectors: undefined,
           detailFilters: {},
+          pageNum: 0,
         },
       },
     };
     store = setupStore(minimalStoreData);
 
-    renderWithProviders(<TemplateCard templateName="template2" />, { store });
+    const onSelect = () => store.dispatch(openQuickViewPanelView());
+
+    renderWithProviders(<TemplateCard templateName="template2" onSelect={onSelect} />, { store });
 
     expect(screen.getByText('Template 2')).toBeDefined();
     screen.getByText('Template 2').click();
