+++ libs/designer/src/lib/ui/templates/cards/templateCard.less
@@ -2,10 +2,14 @@
     min-width: 250px;
     width: 340px;
     border: 1px solid rgba(204, 204, 204, 0.50);
-    border-radius: 2px;
+    border-radius: 8px;
 
     height: 220px;
     box-shadow: 0px 1.6px 3.6px 0px rgba(0, 0, 0, 0.13), 0px 0.3px 0.9px 0px rgba(0, 0, 0, 0.11);
+}
+
+.msla-agent-template-card {
+    height: 150px;
 }
 
 .msla-blank-template-card {
@@ -52,7 +56,6 @@
 
 .msla-template-card-body {
     padding: 0px 12px;
-    height: 180px;
     display: flex;
     flex-direction: column;
     justify-content: center;
