+++ libs/designer/src/lib/ui/templates/filters/templatesearchfilters.tsx
@@ -1,0 +1,275 @@
+import { type FilterObject, TemplatesFilterDropdown } from '@microsoft/designer-ui';
+import type { AppDispatch, RootState } from '../../../core/state/templates/store';
+import { useIntl } from 'react-intl';
+import { useDispatch, useSelector } from 'react-redux';
+import { setConnectorsFilters, setDetailsFilters, setKeywordFilter, setSortKey } from '../../../core/state/templates/manifestSlice';
+import { useEffect, useMemo, useState } from 'react';
+import { getUniqueConnectorsFromConnections } from '../../../core/templates/utils/helper';
+import { useConnector } from '../../../core/state/connection/connectionSelector';
+import { Field, Tab, TabList } from '@fluentui/react-components';
+import { type SelectTabData, type SelectTabEvent, SearchBox, Text, Option, Dropdown } from '@fluentui/react-components';
+import { css } from '@fluentui/utilities';
+
+export type TemplateDetailFilterType = Record<
+  string,
+  {
+    displayName: string;
+    items: FilterObject[];
+  }
+>;
+
+interface GalleryTab {
+  displayName: string;
+  name: string;
+  filterKey?: string;
+}
+
+export interface TemplateSearchAndFilterProps {
+  tabDetails?: GalleryTab[];
+  detailFilters: TemplateDetailFilterType;
+  showFilters?: boolean;
+  searchPlaceholder?: string;
+  cssOverrides?: Record<string, string>;
+}
+
+const templateDefaultTabKey = 'all';
+
+export const TemplateSearchAndFilters = ({
+  tabDetails,
+  searchPlaceholder,
+  showFilters = true,
+  detailFilters,
+  cssOverrides,
+}: TemplateSearchAndFilterProps) => {
+  const dispatch = useDispatch<AppDispatch>();
+  const { sortKey, detailFilters: appliedDetailFilters } = useSelector((state: RootState) => state?.manifest?.filters);
+  const intl = useIntl();
+  const { isConsumption, availableTemplates } = useSelector((state: RootState) => ({
+    isConsumption: state.workflow.isConsumption,
+    availableTemplates: state.manifest.availableTemplates ?? {},
+  }));
+
+  const selectedTabId = appliedDetailFilters?.Type?.[0]?.value ?? templateDefaultTabKey;
+
+  const intlText = {
+    SEARCH: intl.formatMessage({
+      defaultMessage: 'Search',
+      id: 'IUbVFR',
+      description: 'Placeholder text for search templates',
+    }),
+    TYPE: intl.formatMessage({
+      defaultMessage: 'Type',
+      id: 'wfekJ7',
+      description: 'Label text for type filter',
+    }),
+    SORT_BY: intl.formatMessage({
+      defaultMessage: 'Sort By',
+      id: 'ZOIvqN',
+      description: 'Label text for sort by filter',
+    }),
+  };
+
+  const sortOptions = [
+    {
+      key: 'a-to-z',
+      text: intl.formatMessage({
+        defaultMessage: 'A to Z, ascending',
+        id: 'zxF7g+',
+        description: 'Sort by dropdown option of A to Z ascending',
+      }),
+    },
+    {
+      key: 'z-to-a',
+      text: intl.formatMessage({
+        defaultMessage: 'Z to A, descending',
+        id: '1jf3Dq',
+        description: 'Sort by dropdown option of Z to A descending',
+      }),
+    },
+  ];
+
+  const templateTabs = useMemo(() => {
+    const basicTabs: GalleryTab[] = [
+      {
+        name: templateDefaultTabKey,
+        displayName: intl.formatMessage({
+          defaultMessage: 'All',
+          id: 'YX0jQs',
+          description: 'All templates tab',
+        }),
+        filterKey: 'Type',
+      },
+    ];
+
+    if (tabDetails) {
+      return [...basicTabs, ...tabDetails];
+    }
+
+    if (!isConsumption && !!availableTemplates) {
+      basicTabs.push({
+        name: 'Workflow',
+        displayName: intl.formatMessage({
+          defaultMessage: 'Workflows',
+          id: 'fxue5l',
+          description: 'Workflows only templates tab',
+        }),
+        filterKey: 'Type',
+      });
+      basicTabs.push({
+        name: 'Accelerator',
+        displayName: intl.formatMessage({
+          defaultMessage: 'Accelerators',
+          id: 'A5/UwX',
+          description: 'Accelerators only templates tab',
+        }),
+        filterKey: 'Type',
+      });
+    }
+    return basicTabs;
+  }, [intl, tabDetails, isConsumption, availableTemplates]);
+
+  const onTabSelected = (e?: SelectTabEvent, data?: SelectTabData): void => {
+    if (data) {
+      const itemKey = data.value as string;
+      const tabValue = templateTabs.find((tab) => tab.name === itemKey) as GalleryTab;
+      dispatch(
+        setDetailsFilters({
+          filterName: tabValue.filterKey as string,
+          filters:
+            itemKey === templateDefaultTabKey
+              ? undefined
+              : [
+                  {
+                    displayName: itemKey,
+                    value: itemKey,
+                  },
+                ],
+        })
+      );
+    }
+  };
+
+  const placeholderText = searchPlaceholder ?? intlText.SEARCH;
+  return (
+    <div className="msla-templates-search-and-filters">
+      <div className="msla-templates-search-and-sort">
+        <SearchBox
+          className="msla-templates-filters-search-box"
+          placeholder={placeholderText}
+          aria-label={placeholderText}
+          autoFocus={false}
+          onChange={(_e, data) => {
+            dispatch(setKeywordFilter(data.value));
+          }}
+        />
+        <Field className="msla-templates-filters-sort" label={intlText.SORT_BY} orientation="horizontal">
+          <Dropdown
+            className="msla-templates-filters-sort-dropdown"
+            onOptionSelect={(e, option) => dispatch(setSortKey(option.optionValue as string))}
+            value={sortOptions.find((op) => op.key === sortKey)?.text}
+            selectedOptions={[sortKey]}
+            size="small"
+          >
+            {sortOptions.map((op) => (
+              <Option key={op.key} value={op.key}>
+                {op.text}
+              </Option>
+            ))}
+          </Dropdown>
+        </Field>
+      </div>
+
+      {showFilters && <Filters detailFilters={detailFilters} />}
+
+      <div className={css('msla-templates-filters-tabs', cssOverrides?.tabs)}>
+        <TabList selectedValue={selectedTabId} onTabSelect={onTabSelected}>
+          {templateTabs.map((tab) => (
+            <Tab key={tab.name} id={tab.name} value={tab.name}>
+              {tab.displayName}
+            </Tab>
+          ))}
+        </TabList>
+      </div>
+    </div>
+  );
+};
+
+const Filters = ({ detailFilters }: { detailFilters: TemplateDetailFilterType }) => {
+  const dispatch = useDispatch<AppDispatch>();
+  const intl = useIntl();
+  const { isConsumption, availableTemplates, subscriptionId, location } = useSelector((state: RootState) => ({
+    isConsumption: state.workflow.isConsumption,
+    availableTemplates: state.manifest.availableTemplates ?? {},
+    subscriptionId: state.workflow.subscriptionId,
+    location: state.workflow.location,
+  }));
+  const allTemplatesUniqueConnectorIds = useMemo(() => {
+    const skuTemplates = Object.values(availableTemplates).filter((templateManifest) =>
+      templateManifest.skus.includes(isConsumption ? 'consumption' : 'standard')
+    );
+    const allConnections = Object.values(skuTemplates).flatMap((template) => Object.values(template.connections));
+    const uniqueConnectorsFromConnections = getUniqueConnectorsFromConnections(allConnections, subscriptionId, location);
+    return uniqueConnectorsFromConnections.map((connector) => connector.connectorId);
+  }, [availableTemplates, isConsumption, location, subscriptionId]);
+  const [allConnectorsData, setConnectorsData] = useState<Record<string, string>>({});
+  const connectorOptions = useMemo(() => {
+    return allTemplatesUniqueConnectorIds?.map((connectorId) => ({
+      value: connectorId,
+      displayName: allConnectorsData[connectorId] ?? connectorId.split('/').slice(-1)[0],
+    }));
+  }, [allConnectorsData, allTemplatesUniqueConnectorIds]);
+
+  const intlText = {
+    CONNECTORS: intl.formatMessage({
+      defaultMessage: 'Connectors',
+      id: 'KO2eUv',
+      description: 'Label text for connectors filter',
+    }),
+  };
+
+  return (
+    <div className="msla-templates-filters-dropdowns">
+      {allTemplatesUniqueConnectorIds && allTemplatesUniqueConnectorIds.length > 0 && (
+        <TemplatesFilterDropdown
+          filterName={intlText.CONNECTORS}
+          items={connectorOptions}
+          onRenderItem={(item) => (
+            <ConnectorName
+              data={item}
+              setDisplayName={(displayName) => {
+                if (!allConnectorsData[item.key]) {
+                  setConnectorsData({ ...allConnectorsData, [item.key]: displayName });
+                }
+              }}
+            />
+          )}
+          onApplyButtonClick={(filterItems) => {
+            dispatch(setConnectorsFilters(filterItems));
+          }}
+          isSearchable
+        />
+      )}
+      {Object.keys(detailFilters).map((filterName, index) => (
+        <TemplatesFilterDropdown
+          key={index}
+          filterName={detailFilters[filterName].displayName}
+          items={detailFilters[filterName].items}
+          onApplyButtonClick={(filterItems) => {
+            dispatch(setDetailsFilters({ filterName, filters: filterItems }));
+          }}
+        />
+      ))}
+    </div>
+  );
+};
+
+const ConnectorName = ({ data, setDisplayName }: { data: any; setDisplayName: (displayName: string) => void }) => {
+  const { data: connector, isLoading } = useConnector(data.key, /* enabled */ true, /* getCachedData */ true);
+
+  useEffect(() => {
+    if (!isLoading && connector?.properties?.displayName) {
+      setDisplayName(connector.properties.displayName);
+    }
+  }, [connector, isLoading, setDisplayName]);
+  return <Text>{isLoading ? data.key.split('/').slice(-1)[0] : connector?.properties?.displayName}</Text>;
+};
