+++ libs/designer/src/lib/ui/templates/index.tsx
@@ -1,5 +1,6 @@
 export * from './TemplatesDesigner';
-export * from './filters/templateFilters';
+export * from './filters/templatesearchfilters';
 export * from './connections/workflowconnections';
 export * from './templateview';
 export * from './basics/resourcepicker';
+export * from './gallery/templatesgallerywithsearch';
