+++ libs/designer/src/lib/ui/templates/styles.less
@@ -24,23 +24,31 @@
     }
 }
 
-.msla-templates-filters-row {
+.msla-templates-filters-tabs {
     display: flex;
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
 }
 
-.msla-templates-filters {
+.msla-templates-filters-tabs {
+    margin-top: 20px;
+}
+
+.msla-templates-search-and-filters {
     align-items: center;
     justify-content: center;
     width: 100%;
 
-    .msla-templates-filters-search {
-        align-items: center;
-        justify-content: center;
-        width: 585px;
-        margin: auto;
+    .msla-templates-search-and-sort {
+        display: flex;
+        margin: 0 10px 0 10px;
+    }
+
+    .msla-templates-filters-search-box {
+        flex-grow: 2;
+        max-width: 100%;
+        border-radius: 4px;
     }
 
     .msla-templates-filters-dropdowns {
@@ -53,12 +61,12 @@
     .msla-templates-filters-sort {
         display: flex;
         justify-content: flex-end;
-        padding-top: 10px;
+        width: 50%;
+        border-radius: 4px;
     }
 
     .msla-templates-filters-sort-dropdown {
         width: 200px;
-        margin-left: 6px;
     }
 }
